Convert responses into appropriate cell mutations.
