+++ src/extension/ai/converters.ts
@@ -3,6 +3,10 @@
 // See ../vscode_apis.md for an exlanation. It is very helpful for understanding this folder.
 
 import * as vscode from 'vscode'
+import { ServerLifecycleIdentity, getServerConfigurationValue } from '../../utils/configuration'
+import { Serializer } from '../../types'
+import * as serializerTypes from '../grpc/serializerTypes'
+import * as serializer from '../serializer'
 
 // cellToCellData converts a NotebookCell to a NotebookCellData.
 // NotebookCell is an interface used by the editor.
@@ -22,4 +26,35 @@ export function cellToCellData(cell: vscode.NotebookCell): vscode.NotebookCellDa
     data.outputs.push(o)
   }
   return data
+}
+
+// cellProtosToCellData converts an array of RunMe cell protos to an array of VSCode CellData
+export function cellProtosTocellData(cells: serializerTypes.Cell[]): vscode.NotebookCellData[] {
+  let notebook: Serializer.Notebook = {
+    cells: [],
+  }
+  for (let cell of cells) {
+    let kind: vscode.NotebookCellKind = vscode.NotebookCellKind.Markup
+
+    if (cell.kind === serializerTypes.CellKind.CODE) {
+      kind = vscode.NotebookCellKind.Code
+    }
+
+    let newCell: Serializer.Cell = {
+      value: cell.value,
+      metadata: cell.metadata,
+      kind: kind,
+      languageId: cell.languageId,
+      // TODO(jeremy): Should we include outputs? The generate response should never contain outputs so we shouldn't
+      // have to worry about them.
+    }
+    notebook.cells.push(newCell)
+  }
+
+  const identity: ServerLifecycleIdentity = getServerConfigurationValue<ServerLifecycleIdentity>(
+    'lifecycleIdentity',
+    serializerTypes.RunmeIdentity.ALL,
+  )
+  let newCellData = serializer.SerializerBase.revive(notebook, identity)
+  return newCellData
 }
